refactor: template list to template library
