+++ apps/journeys-admin/src/components/TemplateLibrary/TemplateLibrary.tsx
@@ -3,11 +3,13 @@ import Container from '@mui/material/Container'
 import { GetPublishedTemplates_journeys as Journey } from '../../../__generated__/GetPublishedTemplates'
 import { TemplateCard } from '../TemplateCard'
 
-interface TemplateListProps {
+interface TemplateLibraryProps {
   journeys?: Journey[]
 }
 
-export function TemplateList({ journeys }: TemplateListProps): ReactElement {
+export function TemplateLibrary({
+  journeys
+}: TemplateLibraryProps): ReactElement {
   return (
     <Container
       sx={{
