+++ apps/journeys-admin/src/components/TemplateLibrary/index.ts
@@ -1,0 +1,1 @@
+export { TemplateLibrary } from './TemplateLibrary'
