--- apps/journeys-admin/src/components/TemplateList/index.ts
@@ -1,1 +1,0 @@
-export { TemplateList } from './TemplateList'
